Many things changed
Re-Added "- Artist" strings from Main fork of the Rich Presence plugin, changed it so it shows the Album instead, switched around the Title and Album tags aswell as the Format and Song Length because often the Title and Song Length gets cut off. Minor Edits to the About page aswell as some branding changes. Also added a 'project()' to CMakeLists because cmake kept yelling about it.